+++ test/helpers.py
@@ -2,7 +2,7 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-from wampy.async import async_adapter
+from wampy.backends import async_adapter
 
 
 def assert_stops_raising(
