--- wampy/async/errors.py
@@ -1,10 +1,0 @@
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this
-# file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-import eventlet
-import gevent
-
-
-class WampyTimeOut(gevent.Timeout, eventlet.Timeout):
-    pass
